Добавлен модуль компонента модального окна с информацией о фильме
